LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/fill-and-stroke-styles/2d.fillStyle.CSSRGB-expected.txt

2d.fillStyle.CSSRGB

CSSRGB works as color input


FAIL CSSRGB works as color input assert_equals: ctx.fillStyle === '#ff00ff' (got #000000[string], expected #ff00ff[string]) expected "#ff00ff" but got "#000000"
